Add original JsonDoc sources
This does not include maven-specific stuff since we're moving to gradle.
Also, 2 submodules were not imported:

- jsondoc-maven-plugin

The point of Livedoc is to generate a live doc, not deliver a static one, so there is little use for a generated JSON file if we're gonna use the webservice to get the doc anyway.

I'm not interested in carrying on with the maven plugin that generates a static JSON file. Should the need arise later on, these files can still be retrieved via git (or remiplemented from scratch, there is not much, really).

- jsondoc-ui submodule

I don't feel like maintaining this maven-repackageable webapp. Also, the UI will be completely re-done from scratch, so we don't need to bother with an additional subproject to build here.