Remove host in default doc fetch URL
This host is unnecessary because it is equal to the current host, which
is equivalent to not specifying it at all.
If using the webjar, then the doc endpoint in on the same server at
/jsondoc, so "/jsondoc" is sufficient for the fetch to work.
If using an independent UI, then it is likely that the app server URL
ends in /jsondoc, so this string helps anyway because the user just has
to prepend the host.