Prevent rehydration of out-of-date state
The persistence key now contains the Livedoc UI version, so that the state can never be rehydrated from an older version of livedoc.

This is not the best approach as changing often the version of livedoc would result in accumulating data under different keys. However, it is a reasonable compromise while waiting for more info as to how to handle state versions with redux-persist.

Resolves:
https://github.com/joffrey-bion/livedoc/issues/97