Prevent processing of outdated livedoc payload
If the received doc payload was generated by a different version of
Livedoc than the one of the UI, we now prevent the display to avoid
crashes and other instabilities.

In the future, we should prompt the user to know whether he wants to try
a best effort approach, and act accordingly.

Resolves:
https://github.com/joffrey-bion/livedoc/issues/89