Fix flow issues in 3rd party code
After a long read of https://github.com/facebook/flow/issues/869, it appears that the only way of silencing issues in 3rd party code is to ignore the module's file.
This, in turn, leads to "module not found" errors, because flow doesn't load the module anymore.
One quick solution for this was to add a $FlowFixMe comment to silence the errors on the import of the problematic modules.