Add actual bidirectional messages support
Websocket messages were so far stuffed into the same structure as API
operations, which is definitely not ideal. A message is not described by
the same things as an operation and should be treated separately.

The library now uses a different representation and can read more
details about messages. The UI has been adapted to display these
messages separately from the API operations.

Resolves:
https://github.com/joffrey-bion/livedoc/issues/3