Rework merge order for message docs
Every method can yield multiple message docs, because it can potentially handle a message and send multiple others. Also, different methods may use the same message topic/channel and thus the corresponding doc must be merged.

To allow proper override between multiple readers, we need to first merge the message docs of all methods of a given reader, and then override the resulting list with the docs produced by the next reader.

This allows proper handling of a situation where 2 methods declare a @SendTo to the same destination in Spring, and a @MessageChannel annotation with again the same destination is used on either of the methods (or yet another one).
In this case, all 3 annotations must yield a single merged SUBSCRIBE message doc for the destination, and the @MessageChannel must take precedence over the other 2.