Take into account Spring annotation aliases
Spring uses special utilities to read its annotations, because it uses
@AliasFor to alias some annotation attributes and even override
attributes of meta-annotation via synthesized annotations.
Instead of re-implementing this, we can simplify the code and use their
own utilities.

Doing this actually fixes a couple issues that might not have been
noticed before, regarding @GetMapping and the likes, overriding the
attributes of @RequestMapping. In particular, the produces/consumes
attribute are now taken into account.

Resolves:
https://github.com/joffrey-bion/livedoc/issues/107